backend/requirements.txt
asgiref==3.9.1
block-disposable-email==2.0.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
defusedxml==0.7.1
Django==4.2.11
django-admin-rangefilter==0.13.2
django-cors-headers==4.7.0
django-disposable-email-checker==2.0.4
django-redis==6.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
idna==3.10
oauthlib==3.3.1
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python3-openid==3.2.0
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.7.0
sqlparse==0.5.3
urllib3==2.5.0
